WP update
Update various files for enhancements and bug fixes, including adjustments to CSS for better responsiveness, updates to the oEmbed provider for Screencast, and improvements in the comment template to support a new 'novalidate' argument. Additionally, the version number has been bumped to 6.8.2, and security updates have been applied to the ping sites configuration.